backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1